refactor(org.example): major refactoring to support access to approrpriate classes
