+++ src/main/java/org/example/entities/Enemy.java
@@ -1,10 +1,6 @@
 package org.example.entities;
 
-interface AttackBehavior {
-    void attack(Player player);
-    int getDamage();
-    String toString();
-}
+import org.example.entities.behaviors.*;
 
 public abstract class Enemy {
     private final String name;
@@ -55,52 +51,6 @@ public abstract class Enemy {
 
     @Override
     public String toString() {
-        return name + " (" + health + ") " + attackBehavior;
-    }
-}
-
-// --- ENEMIES --------------------------------------------------------------------
-// --- Add only enemies inside this scope -----------------------------------------
-// ----START OF SCOPE--------------------------------------------------------------
-
-// Goblin
-class Goblin extends Enemy {
-    public Goblin() {
-        super("Goblin", 30, new BasicAttack(5));
-    }
-}
-
-// Troll
-class Troll extends Enemy {
-    public Troll() {
-        super("Troll", 50, new BasicAttack(10));
-    }
-}
-
-// ----END OF SCOPE----------------------------------------------------------------
-
-class BasicAttack implements AttackBehavior {
-    private final int damage;
-    
-    public BasicAttack(int damage) {
-        if (damage < 0) {
-            throw new IllegalArgumentException("Damage cannot be negative");
-        }
-        this.damage = damage;
-    }
-
-    @Override
-    public void attack(Player player) {
-        // For later implementation
-    }
-
-    @Override
-    public int getDamage() {
-        return damage;
-    }
-
-    @Override
-    public String toString() {
-        return "+" + damage + " damage";
+        return name + " (" + health + " health, " + attackBehavior + ")";
     }
 }
