+++ src/main/java/org/example/entities/behaviors/AttackBehavior.java
@@ -1,0 +1,9 @@
+package org.example.entities.behaviors;
+
+import org.example.entities.Player;
+
+public interface AttackBehavior {
+    void attack(Player player);
+    int getDamage();
+    String toString();
+}
