+++ src/main/java/org/example/entities/behaviors/CombatBehaviour.java
@@ -1,0 +1,9 @@
+package org.example.entities.behaviors;
+
+import org.example.entities.Enemy;
+
+public interface CombatBehaviour {
+    void attack(Enemy enemy);
+    int getDamage();
+    String toString();
+}
