+++ src/main/java/org/example/entities/behaviors/Effect.java
@@ -1,0 +1,9 @@
+package org.example.entities.behaviors;
+
+import org.example.entities.Player;
+
+public interface Effect {
+    void apply(Player player);
+    int getValue();
+    String toString();
+}
