+++ src/main/java/org/example/entities/items/ItemType.java
@@ -1,4 +1,4 @@
-package org.example.entities;
+package org.example.entities.items;
 
 public enum ItemType {
     WEAPON,
